@@ -1,40 +1,41 @@
 {
-  "name": "gatsby-starter-default",
+  "name": "dor-studio-architecture",
   "private": true,
-  "description": "A simple starter to get up and developing quickly with Gatsby",
+  "description": "Data driven portfolio website for female owned small business architecture firm.",
   "version": "0.1.0",
-  "author": "Kyle Mathews <mathews.kyle@gmail.com>",
+  "author": "Hunter Evanoff <hunter@evanoff.dev>",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "framer-motion": "^4.1.17",
-    "gatsby": "^3.10.2",
-    "gatsby-plugin-gatsby-cloud": "^2.10.1",
-    "gatsby-plugin-image": "^1.10.1",
-    "gatsby-plugin-manifest": "^3.10.0",
-    "gatsby-plugin-offline": "^4.10.0",
-    "gatsby-plugin-react-helmet": "^4.10.0",
-    "gatsby-plugin-react-leaflet": "^3.0.2",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.9.0",
+    "framer-motion": "^6.4.3",
+    "gatsby": "^4.18.0",
+    "gatsby-plugin-gatsby-cloud": "^4.18.1",
+    "gatsby-plugin-image": "^2.18.1",
+    "gatsby-plugin-manifest": "^4.18.1",
+    "gatsby-plugin-material-ui": "^4.1.0",
+    "gatsby-plugin-offline": "^5.18.1",
+    "gatsby-plugin-react-helmet": "^5.18.0",
     "gatsby-plugin-react-svg": "^3.0.1",
-    "gatsby-plugin-sass": "^4.11.0",
-    "gatsby-plugin-sharp": "^3.10.2",
-    "gatsby-source-filesystem": "^3.10.0",
-    "gatsby-transformer-json": "^3.11.0",
-    "gatsby-transformer-sharp": "^3.10.0",
+    "gatsby-plugin-sass": "^5.18.1",
+    "gatsby-plugin-sharp": "^4.18.1",
+    "gatsby-source-filesystem": "^4.18.1",
+    "gatsby-transformer-json": "^4.18.0",
+    "gatsby-transformer-sharp": "^4.18.1",
     "gsap": "^3.7.1",
     "isotope-layout": "^3.0.6",
     "leaflet": "^1.7.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-leaflet": "^3.2.1",
-    "react-slick": "^0.28.1",
+    "react-leaflet": "^4.0.1",
+    "react-slick": "^0.29.0",
     "sass": "^1.37.5",
     "slick-carousel": "^1.8.1"
   },
   "devDependencies": {
-    "gatsby-plugin-material-ui": "^3.0.1",
-    "prettier": "2.3.2"
+    "prettier": "2.7.1"
   },
   "keywords": [
     "gatsby"
