@@ -45,22 +45,22 @@ module.exports = {
         }
       }
     },
-    {
-      resolve: `gatsby-plugin-material-ui`,
-      options: {
-        stylesProvider: {
-          injectFirst: true,
-        },
-      },
-    },
+    // {
+    //   resolve: `gatsby-plugin-material-ui`,
+    //   options: {
+    //     stylesProvider: {
+    //       injectFirst: true,
+    //     },
+    //   },
+    // },
     `custom-mui-theme`,
     `gatsby-plugin-gatsby-cloud`,
-    {
-      resolve: `gatsby-plugin-react-leaflet`,
-      options: {
-        linkStyles: true
-      }
-    }
+    // {
+    //   resolve: `gatsby-plugin-react-leaflet`,
+    //   options: {
+    //     linkStyles: true
+    //   }
+    // }
     // this (optional) plugin enables Progressive Web App + Offline functionality
     // To learn more, visit: https://gatsby.dev/offline
     // `gatsby-plugin-offline`,
