@@ -1,1 +1,7 @@
-export { wrapRootElement } from "./wrapRootElement"
+/* eslint-disable import/prefer-default-export */
+import * as React from 'react';
+import ThemeWrapper from './ThemeWrapper';
+
+export const wrapRootElement = ({ element }) => {
+  return <ThemeWrapper>{element}</ThemeWrapper>;
+};
