@@ -1,14 +1,74 @@
 import React, { useEffect } from "react"
+import { styled } from '@mui/material/styles';
 import L from 'leaflet'
 import { MapContainer, TileLayer, Marker, Tooltip } from "react-leaflet"
 import { useStaticQuery, graphql } from "gatsby"
-import { makeStyles } from "@material-ui/styles"
 import Layout from "../components/layout/layout"
 import Seo from "../components/seo"
 import "leaflet/dist/leaflet.css"
 import icon from 'leaflet/dist/images/marker-icon.png';
 import iconShadow from 'leaflet/dist/images/marker-shadow.png';
 import iconRetina from 'leaflet/dist/images/marker-icon-2x.png';
+
+const PREFIX = 'IndexPage';
+
+const classes = {
+  contactWrapper: `${PREFIX}-contactWrapper`,
+  mapContainer: `${PREFIX}-mapContainer`,
+  textWrapper: `${PREFIX}-textWrapper`,
+  section: `${PREFIX}-section`
+};
+
+const Root = styled('div')((
+  {
+    theme
+  }
+) => ({
+  height: '37.5rem',
+  width: '100%',
+  position: 'relative',
+  '@media(max-width: 47.9375rem)': {
+    height: 'auto'
+  },
+
+  [`& .${classes.mapContainer}`]: {
+    height: '100%',
+    minHeight: '21.875rem'
+  },
+
+  [`& .${classes.textWrapper}`]: {
+    height: '100%',
+    width: '50%',
+    background: 'linear-gradient(to right, rgba(255,255,255,1) 30%, rgba(255,255,255,0) 100%)',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    bottom: 0,
+    padding: '6.25rem',
+    zIndex: 400,
+    color: theme.palette.lightText,
+    '@media(max-width: 75rem)': {
+      padding: '1.875rem',
+    },
+    '@media(max-width: 47.9375rem)': {
+      height: 'auto',
+      position: 'relative',
+      width: '100%',
+      background: 'none'
+    }
+  },
+
+  [`& .${classes.section}`]: {
+    '& p': {
+      margin: 0
+    },
+    textTransform: 'none',
+    marginBottom: '1.3125rem',
+    '&:last-child': {
+      marginBottom: 0
+    }
+  }
+}));
 
 if (typeof window !== 'undefined') {
   let DefaultIcon = L.icon({
@@ -39,7 +99,7 @@ const IndexPage = props => {
       }
     }
   `)
-  const classes = contactStyles(props)
+
   useEffect(() => {
     setTimeout(() => window.dispatchEvent(new Event("resize")), 350)
   }, [])
@@ -57,7 +117,7 @@ const IndexPage = props => {
   }
   if (typeof window !== 'undefined') {
     return (
-      <div className={classes.contactWrapper}>
+      <Root className={classes.contactWrapper}>
         <Seo title="Contact" />
         <MapContainer
           {...mapProps}
@@ -90,56 +150,11 @@ const IndexPage = props => {
             </div>
           ))}
         </div>
-      </div>
-    )
+      </Root>
+    );
   }
   return null
 }
-const contactStyles = makeStyles(theme => ({
-  contactWrapper: { 
-    height: '37.5rem',
-    width: '100%',
-    position: 'relative',
-    '@media(max-width: 47.9375rem)': {
-      height: 'auto'
-    }
-  },
-  mapContainer: {
-    height: '100%',
-    minHeight: '21.875rem'
-  },
-  textWrapper: {
-    height: '100%',
-    width: '50%',
-    background: 'linear-gradient(to right, rgba(255,255,255,1) 30%, rgba(255,255,255,0) 100%)',
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    bottom: 0,
-    padding: '6.25rem',
-    zIndex: 400,
-    color: theme.palette.lightText,
-    '@media(max-width: 75rem)': {
-      padding: '1.875rem',
-    },
-    '@media(max-width: 47.9375rem)': {
-      height: 'auto',
-      position: 'relative',
-      width: '100%',
-      background: 'none'
-    }
-  },
-  section: {
-    '& p': {
-      margin: 0
-    },
-    textTransform: 'none',
-    marginBottom: '1.3125rem',
-    '&:last-child': {
-      marginBottom: 0
-    }
-  },
 
-}))
 IndexPage.Layout = Layout
 export default IndexPage
