@@ -1,13 +1,63 @@
 import React from "react"
+import { styled } from '@mui/material/styles';
 import Layout from "../components/layout/layout"
 import Seo from "../components/seo"
 import { useStaticQuery, graphql } from "gatsby"
 import Subheader from "../components/subheader"
-import { makeStyles } from "@material-ui/core"
 import { GatsbyImage } from "gatsby-plugin-image"
 
-const IndexPage = ({ route, ...props }) => {
-  const classes = servicesStyles(props)
+const PREFIX = 'IndexPage';
+
+const classes = {
+  grid: `${PREFIX}-grid`,
+  gridItem: `${PREFIX}-gridItem`,
+  iconContainer: `${PREFIX}-iconContainer`,
+  textContanier: `${PREFIX}-textContanier`
+};
+
+const Root = styled('React.Fragment')((
+  {
+    theme
+  }
+) => ({
+  [`& .${classes.grid}`]: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr 1fr",
+    gridAutoFlow: "column",
+    gridTemplateRows: "repeat(2, auto)",
+    gap: "1.875rem",
+    width: "100%",
+    textTransform: "none",
+    marginBottom: "2.1875rem",
+    "@media(max-width: 47.9375rem)": {
+      gridTemplateColumns: "none",
+      gridTemplateRows: "none",
+      gridAutoFlow: "unset",
+    },
+  },
+
+  [`& .${classes.gridItem}`]: {
+    padding: "1.875rem",
+    backgroundColor: theme.palette.gridItem,
+  },
+
+  [`& .${classes.iconContainer}`]: {
+    width: "100%",
+    textAlign: "center",
+    color: theme.palette.primary.main,
+    fontSize: "2.5em",
+    padding: "0 0.625rem",
+    lineHeight: "3.125rem",
+  },
+
+  [`& .${classes.textContanier}`]: {
+    width: "100%",
+    padding: "0 0.625rem",
+  }
+}));
+
+const IndexPage = ({ route }) => {
+
   const data = useStaticQuery(graphql`
     {
       servicesJson {
@@ -38,7 +88,7 @@ const IndexPage = ({ route, ...props }) => {
   const servicesData = data.servicesJson
 
   return (
-    <>
+    <Root>
       <Seo title="Services" />
       <Subheader subheader={servicesData.subheader} />
       <div className={classes.grid}>
@@ -81,43 +131,9 @@ const IndexPage = ({ route, ...props }) => {
           </React.Fragment>
         ))}
       </div>
-    </>
-  )
+    </Root>
+  );
 }
-
-const servicesStyles = makeStyles(theme => ({
-  grid: {
-    display: "grid",
-    gridTemplateColumns: "1fr 1fr 1fr",
-    gridAutoFlow: "column",
-    gridTemplateRows: "repeat(2, auto)",
-    gap: "1.875rem",
-    width: "100%",
-    textTransform: "none",
-    marginBottom: "2.1875rem",
-    "@media(max-width: 47.9375rem)": {
-      gridTemplateColumns: "none",
-      gridTemplateRows: "none",
-      gridAutoFlow: "unset",
-    },
-  },
-  gridItem: {
-    padding: "1.875rem",
-    backgroundColor: theme.palette.gridItem,
-  },
-  iconContainer: {
-    width: "100%",
-    textAlign: "center",
-    color: theme.palette.primary.main,
-    fontSize: "2.5em",
-    padding: "0 0.625rem",
-    lineHeight: "3.125rem",
-  },
-  textContanier: {
-    width: "100%",
-    padding: "0 0.625rem",
-  },
-}))
 
 IndexPage.Layout = Layout
 export default IndexPage
